pyproject.toml
[tool.poetry]
name = "picpick"
version = "0.1.0"
description = ""
authors = ["michaeldel <michaeldel@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^7.1.1"
click = "^7.1.2"
bidict = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^19.10b0", allow-prereleases = true}
flake8 = "^3.7.9"
mypy = "^0.770"

[tool.black]
target-version = ['py38']

line-length = 88
skip-string-normalization = true

include = '\.pyi?$'
exclude = '\/\..+$'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

